Add the list of container read only file types
These are the types of files that container domains are allowed to read.
We want this for contianer storage, to allow us to do some deduplicating
of SELinux types for use within container/storage.

This list is very static and has been basically the same for years.
Generating this programatically, while possible is very costly.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>